pyproject.toml
[project]
name = "telegram-gedam-a-studio"
version = "0.1.0"
description = ""
authors = [
    {name = "21dagi",email = "21dagmawinegussie@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram (==3.2.0)",
    "python-dotenv (==1.0.0)",
    "firebase-admin (==6.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
